Add basic email ticket notifications
The book didn't mention about adding the email_spec gem to the
spec_helper file. Check out the project's Github page and they mentioned
that so that's how I got the gem to work.
Had to set the default_url_options for the ActionMailer in test env
config file for the  mailer to send.